# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Capacity & Spend Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      parameters:
        - { name: skill, in: query, schema: { type: string } }
        - { name: location, in: query, schema: { type: string } }
        - { name: employmentType, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Resource" }
    post:
      operationId: createResource
      tags: [resources]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResourceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resource" }
  /resources/{id}:
    get:
      operationId: getResource
      tags: [resources]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resource" }
    patch:
      operationId: updateResource
      tags: [resources]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResourceInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resource" }
    delete:
      operationId: deleteResource
      tags: [resources]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Resource:
      type: object
      required:
        - id
        - name
        - role
        - primarySkill
        - secondarySkills
        - location
        - rate
        - rateType
        - currency
        - fxRateToUsd
        - employmentType
        - availabilityPct
        - costCentre
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        primarySkill: { type: string }
        secondarySkills:
          type: array
          items: { type: string }
        location: { type: string }
        rate: { type: number }
        rateType: { type: string, enum: [daily, monthly] }
        currency: { type: string }
        fxRateToUsd: { type: number }
        employmentType: { type: string, enum: [FTE, Contractor, Vendor] }
        availabilityPct: { type: number }
        costCentre: { type: string }
        functionId: { type: integer, nullable: true }
        functionName: { type: string, nullable: true }
    ResourceInput:
      type: object
      required:
        - name
        - role
        - primarySkill
        - secondarySkills
        - location
        - rate
        - rateType
        - currency
        - employmentType
        - availabilityPct
        - costCentre
      properties:
        name: { type: string }
        role: { type: string }
        primarySkill: { type: string }
        secondarySkills:
          type: array
          items: { type: string }
        location: { type: string }
        rate: { type: number }
        rateType: { type: string, enum: [daily, monthly] }
        currency: { type: string }
        fxRateToUsd: { type: number }
        employmentType: { type: string, enum: [FTE, Contractor, Vendor] }
        availabilityPct: { type: number }
        costCentre: { type: string }
        functionId: { type: integer, nullable: true }
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [Admin, "PMO Lead", "Project Manager", "Finance Controller", "Resource Manager", "Demand Owner", Viewer]
        createdAt: { type: string, format: date-time }
        functionId: { type: integer, nullable: true }
        functionName: { type: string, nullable: true }
    CreateUserInput:
      type: object
      required: [email, password, name, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }
        role:
          type: string
          enum: [Admin, "PMO Lead", "Project Manager", "Finance Controller", "Resource Manager", "Demand Owner", Viewer]
        functionId: { type: integer, nullable: true }
    UpdateUserInput:
      type: object
      properties:
        name: { type: string }
        role:
          type: string
          enum: [Admin, "PMO Lead", "Project Manager", "Finance Controller", "Resource Manager", "Demand Owner", Viewer]
        password: { type: string, minLength: 8 }
        functionId: { type: integer, nullable: true }
...