# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Capacity & Spend Control API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceEmploymentType } from "./resourceEmploymentType";
import type { ResourceRateType } from "./resourceRateType";

export interface Resource {
  id: number;
  name: string;
  role: string;
  primarySkill: string;
  secondarySkills: string[];
  location: string;
  rate: number;
  rateType: ResourceRateType;
  currency: string;
  fxRateToUsd: number;
  employmentType: ResourceEmploymentType;
  availabilityPct: number;
  costCentre: string;
  functionId?: number | null;
  functionName?: string | null;
}