# lib/api-zod/src/generated/types/resourceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Capacity & Spend Control API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceInputEmploymentType } from "./resourceInputEmploymentType";
import type { ResourceInputRateType } from "./resourceInputRateType";

export interface ResourceInput {
  name: string;
  role: string;
  primarySkill: string;
  secondarySkills: string[];
  location: string;
  rate: number;
  rateType: ResourceInputRateType;
  currency: string;
  fxRateToUsd?: number;
  employmentType: ResourceInputEmploymentType;
  availabilityPct: number;
  costCentre: string;
  functionId?: number | null;
}